+++ src/WPGet_Elementor_Widgets/Modules/Site_Settings_General/Module_Controller.php
@@ -2,6 +2,7 @@
 namespace WPGet_Elementor_Widgets\Modules\Site_Settings_General;
 
 use Elementor\Core\Kits\Documents\Kit;
+use Elementor\Plugin;
 use WPGet_Elementor_Widgets\Helper\ElementorHelper;
 
 class Module_Controller
@@ -18,6 +19,7 @@ class Module_Controller
     public function __construct()
     {
         add_action('init', [$this, 'wp_init']);
+        add_action( 'elementor/editor/footer', [$this, 'print_settings']);
     }
 
     public function wp_init(){
@@ -29,6 +31,22 @@ class Module_Controller
     public function init_site_settings(Kit $kit){
         $kit->register_tab( 'settings-wpget', Site_Settings::class );
     }
+
+    public function print_settings()
+    {
+        $kit = Plugin::$instance->kits_manager->get_active_kit();
+        $brand_icon = $kit->get_settings('branding_icon');
+        if(!empty($brand_icon['url'])):
+        ?>
+<style>
+    :root{
+        --wpget-svg-icon: url(<?php echo $brand_icon['url'] ?>);
+    }
+</style>
+<?php
+endif;
+    }
+
 
 }